backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
openai>=1.50
google-generativeai>=0.8
tavily-python>=0.5
python-dotenv>=1.0
pydantic>=2.0
httpx>=0.27
